Use play_turn_fast() to optimize the Monte Carlo AI
Now, this isn't a pure optimization, since it also changes the behavior
of the AI. Because it's fast, I decided to use longer playouts. This may
also help its intelligence by making the evaluation more consistent
(hopefully, it usually simulates until the end of the game rather than
usually cutting off with a naive evaluation).