Removed the conrod UI because it was broken
I came back after a long break to find that I was getting a segfault
somewhere in glium. In order to make it work for the moment, I just
removed the UI.

Conrod wasn't easy to use anyway, so if and when I make another GUI,
it'll probably use a web front-end instead. Still, I'm putting this in a
separate branch because it's a regression.